+++ src/main/java/com/ats/tril/model/indent/DashIndentDetails.java
@@ -17,7 +17,7 @@ public class DashIndentDetails implements Serializable{
 	
 	private String itemCode;
 	
-	private int indQty;
+	private float indQty;
 	
 	private String indItemUom;
 	
@@ -33,7 +33,7 @@ public class DashIndentDetails implements Serializable{
 	
 	private int indDStatus;
 	
-	private int	indFyr;
+	private float	indFyr;
 
 	public int getIndDId() {
 		return indDId;
@@ -83,11 +83,11 @@ public class DashIndentDetails implements Serializable{
 		this.itemCode = itemCode;
 	}
 
-	public int getIndQty() {
+	public float getIndQty() {
 		return indQty;
 	}
 
-	public void setIndQty(int indQty) {
+	public void setIndQty(float indQty) {
 		this.indQty = indQty;
 	}
 
@@ -147,11 +147,11 @@ public class DashIndentDetails implements Serializable{
 		this.indDStatus = indDStatus;
 	}
 
-	public int getIndFyr() {
+	public float getIndFyr() {
 		return indFyr;
 	}
 
-	public void setIndFyr(int indFyr) {
+	public void setIndFyr(float indFyr) {
 		this.indFyr = indFyr;
 	}
